+++ stac_fastapi/extensions/stac_fastapi/extensions/core/pagination/__init__.py
@@ -1,0 +1,6 @@
+"""pagination classes as extensions."""
+
+from .pagination import PaginationExtension
+from .token_pagination import TokenPaginationExtension
+
+__all__ = ["PaginationExtension", "TokenPaginationExtension"]