+++ stac_fastapi/pgstac/stac_fastapi/pgstac/app.py
@@ -2,37 +2,44 @@
 from fastapi.responses import ORJSONResponse
 
 from stac_fastapi.api.app import StacApi
+from stac_fastapi.api.models import create_get_request_model, create_post_request_model
 from stac_fastapi.extensions.core import (
     ContextExtension,
     FieldsExtension,
-    QueryExtension,
     SortExtension,
+    TokenPaginationExtension,
     TransactionExtension,
 )
 from stac_fastapi.pgstac.config import Settings
 from stac_fastapi.pgstac.core import CoreCrudClient
 from stac_fastapi.pgstac.db import close_db_connection, connect_to_db
+from stac_fastapi.pgstac.extensions import QueryExtension
 from stac_fastapi.pgstac.transactions import TransactionsClient
 from stac_fastapi.pgstac.types.search import PgstacSearch
 
 settings = Settings()
+extensions = [
+    TransactionExtension(
+        client=TransactionsClient(),
+        settings=settings,
+        response_class=ORJSONResponse,
+    ),
+    QueryExtension(),
+    SortExtension(),
+    FieldsExtension(),
+    TokenPaginationExtension(),
+    ContextExtension(),
+]
+
+post_request_model = create_post_request_model(extensions, base_model=PgstacSearch)
 
 api = StacApi(
     settings=settings,
-    extensions=[
-        TransactionExtension(
-            client=TransactionsClient(),
-            settings=settings,
-            response_class=ORJSONResponse,
-        ),
-        QueryExtension(),
-        SortExtension(),
-        FieldsExtension(),
-        ContextExtension(),
-    ],
-    client=CoreCrudClient(),
-    search_request_model=PgstacSearch,
+    extensions=extensions,
+    client=CoreCrudClient(post_request_model=post_request_model),
     response_class=ORJSONResponse,
+    search_get_request_model=create_get_request_model(extensions),
+    search_post_request_model=post_request_model,
 )
 app = api.app