+++ stac_fastapi/api/stac_fastapi/api/config.py
@@ -21,5 +21,4 @@ class ApiExtensions(enum.Enum):
 class AddOns(enum.Enum):
     """Enumeration of available third party add ons."""
 
-    tiles = "tiles"
     bulk_transaction = "bulk-transaction"