+++ stac_fastapi/extensions/stac_fastapi/extensions/core/__init__.py
@@ -1,9 +1,8 @@
 """stac_api.extensions.core module."""
-from stac_fastapi.extensions.third_party.tiles import TilesExtension
-
 from .context import ContextExtension
 from .fields import FieldsExtension
 from .filter import FilterExtension
+from .pagination import PaginationExtension, TokenPaginationExtension
 from .query import QueryExtension
 from .sort import SortExtension
 from .transaction import TransactionExtension
@@ -12,8 +11,9 @@ __all__ = (
     "ContextExtension",
     "FieldsExtension",
     "FilterExtension",
+    "PaginationExtension",
     "QueryExtension",
     "SortExtension",
-    "TilesExtension",
+    "TokenPaginationExtension",
     "TransactionExtension",
 )