+++ stac_fastapi/extensions/stac_fastapi/extensions/core/fields/request.py
@@ -1,0 +1,71 @@
+"""Request models for the fields extension."""
+
+from typing import Dict, Optional, Set
+
+import attr
+from pydantic import BaseModel, Field
+
+from stac_fastapi.types.config import Settings
+from stac_fastapi.types.search import APIRequest, str2list
+
+
+class PostFieldsExtension(BaseModel):
+    """FieldsExtension.
+
+    Attributes:
+        include: set of fields to include.
+        exclude: set of fields to exclude.
+    """
+
+    include: Optional[Set[str]] = set()
+    exclude: Optional[Set[str]] = set()
+
+    @staticmethod
+    def _get_field_dict(fields: Optional[Set[str]]) -> Dict:
+        """Pydantic include/excludes notation.
+
+        Internal method to create a dictionary for advanced include or exclude of pydantic fields on model export
+        Ref: https://pydantic-docs.helpmanual.io/usage/exporting_models/#advanced-include-and-exclude
+        """
+        field_dict = {}
+        for field in fields or []:
+            if "." in field:
+                parent, key = field.split(".")
+                if parent not in field_dict:
+                    field_dict[parent] = {key}
+                else:
+                    field_dict[parent].add(key)
+            else:
+                field_dict[field] = ...  # type:ignore
+        return field_dict
+
+    @property
+    def filter_fields(self) -> Dict:
+        """Create pydantic include/exclude expression.
+
+        Create dictionary of fields to include/exclude on model export based on the included and excluded fields passed
+        to the API
+        Ref: https://pydantic-docs.helpmanual.io/usage/exporting_models/#advanced-include-and-exclude
+        """
+        # Always include default_includes, even if they
+        # exist in the exclude list.
+        include = (self.include or set()) - (self.exclude or set())
+        include |= Settings.get().default_includes or set()
+
+        return {
+            "include": self._get_field_dict(include),
+            "exclude": self._get_field_dict(self.exclude),
+        }
+
+
+@attr.s
+class FieldsExtensionGetRequest(APIRequest):
+    """Additional fields for the GET request."""
+
+    fields: Optional[str] = attr.ib(default=None, converter=str2list)
+
+
+class FieldsExtensionPostRequest(BaseModel):
+    """Additional fields and schema for the POST request."""
+
+    fields: Optional[PostFieldsExtension] = Field(PostFieldsExtension())