+++ CHANGES.md
@@ -10,6 +10,7 @@
 
 ### Fixed
 
+* The minimum `limit` value for searches is now 1 ([#296](https://github.com/stac-utils/stac-fastapi/pull/296))
 * Links stored with Collections and Items (e.g. license links) are now returned with those STAC objects ([#282](https://github.com/stac-utils/stac-fastapi/pull/282))
 
 ## [2.2.0]